requirements.txt
flask==2.3.3
apscheduler==3.10.1
schedule==1.2.1
instagrapi==1.18.4
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.2.0
gnews==0.3.6
python-dotenv==1.0.0
